+++ lib/liquid/block_body.rb
@@ -45,7 +45,7 @@ module Liquid
           end
           tag_name = Regexp.last_match(1)
           markup   = Regexp.last_match(2)
-          unless (tag = registered_tags[tag_name])
+          unless (tag = registered_tag(tag_name, parse_context))
             # end parsing if we reach an unknown tag and let the caller decide
             # determine how to proceed
             return yield tag_name, markup
@@ -132,7 +132,7 @@ module Liquid
             next
           end
 
-          unless (tag = registered_tags[tag_name])
+          unless (tag = registered_tag(tag_name, parse_context))
             # end parsing if we reach an unknown tag and let the caller decide
             # determine how to proceed
             return yield tag_name, markup
@@ -248,8 +248,9 @@ module Liquid
       BlockBody.raise_missing_variable_terminator(token, parse_context)
     end
 
-    def registered_tags
-      Template.tags
+    def registered_tag(tag_name, parse_context)
+      context_tag = parse_context.tags[tag_name] if parse_context.tags.respond_to?(:[])
+      context_tag || Template.tags[tag_name]
     end
   end
 end